% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmailr.R
\docType{package}
\name{retinagmailr}
\alias{retinagmailr}
\alias{retinagmailr-package}
\title{\pkg{retinagmailr} makes gmail access easy.}
\description{
\code{gmailr} provides an interface to the gmail api \url{https://developers.google.com/gmail/api/}
}
